better use of the new quote parser
