+++ EmailValidator/Parser/LocalPart.php
@@ -7,14 +7,10 @@ use Egulias\EmailValidator\Result\Result;
 use Egulias\EmailValidator\Result\ValidEmail;
 use Egulias\EmailValidator\Exception\DotAtEnd;
 use Egulias\EmailValidator\Result\InvalidEmail;
-use Egulias\EmailValidator\Warning\CFWSWithFWS;
 use Egulias\EmailValidator\Warning\LocalTooLong;
-use Egulias\EmailValidator\Warning\QuotedString;
-use Egulias\EmailValidator\Exception\ExpectingAT;
 use Egulias\EmailValidator\Exception\ExpectingATEXT;
 use Egulias\EmailValidator\Result\Reason\DotAtStart;
 use Egulias\EmailValidator\Exception\UnopenedComment;
-use Egulias\EmailValidator\Exception\UnclosedQuotedString;
 
 class LocalPart extends Parser
 {
@@ -31,13 +27,7 @@ class LocalPart extends Parser
             }
 
             if ($parseDQuote) {
-                $dquoteParser = new DoubleQuote($this->lexer);
-                $parseDQuote = $dquoteParser->parse($localPart);
-                $warns = $dquoteParser->getWarnings();
-                foreach ($warns as $code => $dWarning) {
-                    $this->warnings[$code] = $dWarning;
-                }
-                //$parseDQuote = $this->parseDoubleQuote();
+                $parseDQuote = $this->parseDoubleQuote();
             }
 
             if ($this->lexer->token['type'] === EmailLexer::S_OPENPARENTHESIS) {
@@ -84,59 +74,13 @@ class LocalPart extends Parser
             return $this->lexer->token['type'] === EmailLexer::S_DOT && null === $this->lexer->getPrevious()['type'];
     }
 
-    /**
-     * @return bool
-     */
-    protected function parseDoubleQuote()
+    protected function parseDoubleQuote() : bool
     {
-        if ($this->lexer->token['type'] !== EmailLexer::S_DQUOTE) {
-            return true;
-        }
-        if(!$this->checkDQUOTE(false)) return false;
-        $parseAgain = true;
-        $special = array(
-            EmailLexer::S_CR => true,
-            EmailLexer::S_HTAB => true,
-            EmailLexer::S_LF => true
-        );
-
-        $invalid = array(
-            EmailLexer::C_NUL => true,
-            EmailLexer::S_HTAB => true,
-            EmailLexer::S_CR => true,
-            EmailLexer::S_LF => true
-        );
-        $setSpecialsWarning = true;
-
-        $this->lexer->moveNext();
-
-        while ($this->lexer->token['type'] !== EmailLexer::S_DQUOTE && null !== $this->lexer->token['type']) {
-            $parseAgain = false;
-            if (isset($special[$this->lexer->token['type']]) && $setSpecialsWarning) {
-                $this->warnings[CFWSWithFWS::CODE] = new CFWSWithFWS();
-                $setSpecialsWarning = false;
-            }
-            if ($this->lexer->token['type'] === EmailLexer::S_BACKSLASH && $this->lexer->isNextToken(EmailLexer::S_DQUOTE)) {
-                $this->lexer->moveNext();
-            }
-
-            $this->lexer->moveNext();
-
-            if (!$this->escaped() && isset($invalid[$this->lexer->token['type']])) {
-                throw new ExpectingATEXT();
-            }
-        }
-
-        $prev = $this->lexer->getPrevious();
-
-        if ($prev['type'] === EmailLexer::S_BACKSLASH) {
-            if (!$this->checkDQUOTE(false)) {
-                throw new UnclosedQuotedString();
-            }
-        }
-
-        if (!$this->lexer->isNextToken(EmailLexer::S_AT) && $prev['type'] !== EmailLexer::S_BACKSLASH) {
-            throw new ExpectingAT();
+        $dquoteParser = new DoubleQuote($this->lexer);
+        $parseAgain = $dquoteParser->parse("remove useless arg");
+        $warns = $dquoteParser->getWarnings();
+        foreach ($warns as $code => $dWarning) {
+            $this->warnings[$code] = $dWarning;
         }
 
         return $parseAgain;
